Separating gemm implementations (ref, local mem, no local mem) (#163)
* Separating gemm implementations
